requirements.txt
numpy 
streamlit 
python-dotenv 
pandas 
scikit-learn
pyttsx3 
google-generativeai

